+++ Include/structs.h
@@ -10,13 +10,13 @@
  * */
 /*
 typedef struct s_color {
-	float red, green, blue;
+	double red, green, blue;
 } colour;
 
 
 typedef struct s_material{
 	colour diffuse;
-	float	reflection;
+	double	reflection;
 }	t_material;
 
 typedef struct	s_light {
@@ -27,7 +27,7 @@ typedef struct	s_light {
 
 typedef struct	s_sphere {
 	t_point	pos;
-	float	radius;
+	double	radius;
 	int		material;
 }	t_sphere;
 */
@@ -71,9 +71,9 @@ typedef struct _data {
 
 typedef struct	point {
 
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_point;
 
     // t_point left_corner;
@@ -83,17 +83,17 @@ typedef struct	s_ray {
 
     t_point origin;
     t_point direction;
-    float   t;
+    double   t;
 }	t_ray;
 
 typedef struct matrix
 {
-	float	M[4][4];
+	double	M[4][4];
 }				t_matrix;
 
 typedef struct _Ambient_light
 {
-    float   ratio;//ambient lighting ratio in range [0.0,1.0]:
+    double   ratio;//ambient lighting ratio in range [0.0,1.0]:
     size_t  color[3];//R,G,B colors in range [0-255]: 255, 255, 255
     size_t  repetition;
 } t_Ambient;
@@ -110,13 +110,13 @@ typedef struct _Camera
 typedef struct _Light
 {
     t_point   cordinates;//x,y,z coordinates of the view
-    float    brightenss;//the light brightness ratio in range [0.0,1.0]
+    double    brightenss;//the light brightness ratio in range [0.0,1.0]
     size_t   color[3];//R,G,B colors in range [0-255]: 255, 255, 255
     size_t   repetition;
 } t_Light;
 
 typedef struct s_intersect {
-	float	tmin;
+	double	tmin;
 	t_point	closestPoint;
 	size_t	*color;
 
@@ -136,7 +136,7 @@ typedef struct _Plane
 typedef struct _Sphere
 {
 	t_point         center;//x,y,z coordinates of the view
-	float			radius;
+	double			radius;
 	size_t			color[3];//R,G,B colors in range [0-255]: 255, 255, 255
 	size_t			repetition;
 
@@ -148,8 +148,8 @@ typedef struct _Cylinder
     t_point             cordinates;//x,y,z coordinates of the view
     t_point             orientation;//3d normalized orientation vector. In range [-1,1]
     size_t              color[3];//R,G,B colors in range [0-255]: 255, 255, 255
-    float               redius;
-    float               height;
+    double               redius;
+    double               height;
     size_t              repetition;
     // struct _Cylinder    *next;
 } t_Cylinder;
